+++ packages/iov-keycontrol/src/userprofile.ts
@@ -45,9 +45,13 @@ const passwordHashingOptionsVersion2: Argon2idOptions = {
 // Must be 16 bytes due to implementation limitations.
 const userProfileSalt = toAscii("core-userprofile");
 
-// the format version in which profiles are stored
-// TODO: activate version 2 in 0.15 (https://github.com/iov-one/iov-core/issues/1025)
-const latestFormatVersion = 1;
+/**
+ * The format version in which profiles are stored
+ *
+ * When downgrading IOV-Core should be possible, read support for a new format version
+ * should be rolled out before this number is bumped.
+ */
+const storageFormatVersion = 2;
 
 export interface UserProfileEncryptionKey {
   readonly formatVersion: number;
@@ -96,7 +100,7 @@ export class UserProfile {
    */
   public static async deriveEncryptionKey(
     password: string,
-    formatVersion: number = latestFormatVersion,
+    formatVersion: number = storageFormatVersion,
   ): Promise<UserProfileEncryptionKey> {
     switch (formatVersion) {
       case 1:
@@ -194,8 +198,6 @@ export class UserProfile {
     }
 
     await DatabaseUtils.clear(db);
-
-    const storageFormatVersion = latestFormatVersion;
 
     const key =
       typeof encryptionSecret === "string"